requirements.txt
cffi==1.14.6
code128==0.2b2
cssselect2==0.4.1
docxcompose==1.3.3
docxtpl==0.14.2
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
python-barcode==0.13.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
pyvips==2.1.15
qrcode==7.3.1
reportlab==3.6.1
six==1.16.0
svglib==1.1.0
tinycss2==1.1.0
webencodings==0.5.1
yapf==0.31.0
